pyproject.toml
[tool.poetry]
name = "sendnewvachh"
version = "0.1.0"
description = ""
authors = ["zalim <zalimpshigotizhev@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "13.15"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
